Move main code in json dir; more windows-friendly.
The motivation for moving things into the json directory is to enable a
universal install procedure like this:

* Copy the cstructs and json directories into your project.

aaaand, that's it.

This is nice because it works on Xcode, on visual studio, and with
makefiles. It is also compatible with anything else depending on
cstructs, so you only need one copy of cstructs in your project.

Sadly, the windows-friendly stuff is a tad ugly. I hope to one day
separate this out so we can keep json.{h,c} and jsonutil.{h,c} clean.