Add a trailing space to clarify a comment end.
I have a line that ends with a \ as part of a //-comment.

In my head and in clang, the backslash is a valid ending of such a
comment. However, in Visual Studio, the backslash indicates to continue
the comment on the next line. Adding a trailing space keeps the visual
appearance of the comment while making it work on both compilers.