Replace some nested comments with end-of-line ones
I decided I like end-of-line comments more because they avoid too much mixing of
code and comments. From one point of view, when you don't embed comments in
code, then everything "to the left" is what the compiler sees and what is being
definitively executed, while everything "to the right" is how a human is
thinking about the code. I appreciate the conceptual simplicity of that
separation.